Update DwC button, fix map label for aggregate maps, split nomenclature panels
