+++ src/modules/otus/components/Panel/PanelNomenclature/PanelNomenclature.vue
@@ -1,12 +1,59 @@
 <template>
-  <PanelNomenlcatureCitations :list="store.catalog.timeline" />
-  <PanelNomenclatureReferences :list="store.catalog.sources" />
+  <VCard>
+    <VCardHeader class="flex justify-between">
+      <h2 class="text-md">
+        Nomenclature ({{ store.catalog.timeline.length }})
+      </h2>
+      <PanelDropdown
+        :menu-options="menuOptions"
+        panel-key="taxonomy"
+      />
+    </VCardHeader>
+
+    <ul class="text-sm">
+      <CitationRow
+        v-for="citation in citationList.first"
+        :key="citation.label"
+        :citation="citation"
+      />
+
+      <PanelNomenclatureShowMore
+        v-if="!showAll && citationList.middle.length"
+        :count="citationList.middle.length"
+        @click="showAll = true"
+      />
+    </ul>
+    <AnimationOpacity>
+      <ul
+        class="text-sm"
+        v-show="showAll"
+      >
+        <CitationRow
+          v-for="citation in citationList.middle"
+          :key="citation.label"
+          :citation="citation"
+        />
+      </ul>
+    </AnimationOpacity>
+    <ul class="text-sm">
+      <CitationRow
+        v-for="citation in citationList.last"
+        :key="citation.label"
+        :citation="citation"
+      />
+    </ul>
+  </VCard>
 </template>
 
 <script setup>
+import { ref, computed } from 'vue'
+import { splitList } from './splitList'
 import { useOtuStore } from '@/modules/otus/store/store'
-import PanelNomenlcatureCitations from './PanelNomenclatureCitations.vue'
-import PanelNomenclatureReferences from './PanelNomenclatureReferences.vue'
+import CitationRow from './PanelCitationRow.vue'
+import PanelNomenclatureShowMore from './PanelNomenclatureShowMore.vue'
+import PanelDropdown from '../PanelDropdown.vue'
+
+const MAX_CITATIONS = 2
 
 const props = defineProps({
   otuId: {
@@ -31,4 +78,16 @@ const props = defineProps({
 })
 
 const store = useOtuStore()
+
+const showAll = ref(false)
+const citationList = computed(() =>
+  splitList(store.catalog.timeline, MAX_CITATIONS)
+)
+
+const menuOptions = computed(() => [
+  {
+    label: showAll.value ? 'Show less' : 'Show all',
+    action: () => (showAll.value = !showAll.value)
+  }
+])
 </script>
