+++ src/modules/otus/components/Panel/PanelNomenclatureReferences/PanelReferenceRow.vue
