+++ src/modules/otus/constants/overviewLayout.js
@@ -3,6 +3,7 @@ import PanelGallery from '../components/Panel/PanelGallery/Gallery.vue'
 import PanelTypeSpecimen from '../components/Panel/PanelTypeSpecimen/PanelTypeSpecimen.vue'
 import PanelTypeDesignation from '../components/Panel/PanelTypeDesignation/PanelTypeDesignation.vue'
 import PanelNomenclature from '../components/Panel/PanelNomenclature/PanelNomenclature.vue'
+import PanelNomenclatureReference from '../components/Panel/PanelNomenclatureReferences/PanelNomenclatureReferences.vue'
 import PanelMap from '../components/Panel/PanelMap/PanelMap.vue'
 import PanelDescendants from '../components/Panel/PanelDescendants/Descendants.vue'
 import PanelContent from '../components/Panel/PanelContent/PanelContent.vue'
@@ -19,7 +20,8 @@ export const overviewLayout = {
       component: PanelTypeDesignation,
       available: [FAMILY_GROUP, GENUS_GROUP]
     },
-    { component: PanelNomenclature }
+    { component: PanelNomenclature },
+    { component: PanelNomenclatureReference }
   ],
 
   right: [
